refactor: base.html and trips/views - delete leafletjs from base as it's not necessary for Folium - move utils in view to proper location in utils - refactor tests accordingly
