+++ templates/base.html
@@ -20,9 +20,6 @@
         {% tailwind_css %}
         <!-- Custom CSS -->
         <link rel="stylesheet" href="{% static 'css/custom.css' %}" />
-        <!-- LeafletJS -->
-        <link rel="stylesheet" href="{% static 'css/leaflet.css' %}" />
-        <script src="{% static 'js/leaflet.js' %}"></script>
     </head>
     <body hx-headers='{"X-CSRFToken": "{{ csrf_token }}"}'>
         <section>
